+++ packages/process/test/Main.hs
@@ -1,15 +1,12 @@
 module Main where
 
-import Polysemy.Process.Test.ProcessTest (test_process)
-import Polysemy.Test (unitTest)
+import Polysemy.Process.Test.ProcessTest (test_processAll)
 import Test.Tasty (TestTree, defaultMain, testGroup)
 
 tests :: TestTree
 tests =
   testGroup "main" [
-    testGroup "process" [
-      unitTest "process" test_process
-    ]
+    test_processAll
   ]
 
 main :: IO ()